requirements.txt
Flask
python-dotenv
pymongo[srv]
pdfplumber
google-generativeai